TALLER_05/requirements.txt
flask
flask-cors
python-dotenv
mariadb
